+++ src/config/args/restore_object.rs
@@ -21,7 +21,6 @@ fn parse_tier(s: &str) -> Result<Tier, String> {
 #[derive(Parser, Clone, Debug)]
 pub struct RestoreObjectArgs {
     #[arg(
-        env,
         help = "s3://<BUCKET_NAME>/<KEY>",
         value_parser = storage_path::check_storage_path,
         required_unless_present = "auto_complete_shell"